Snapshot settings new flow
