+++ kotsadm/web/src/scss/utilities/icons.scss
@@ -596,24 +596,24 @@
   height: 16px;
 }
 .icon.awsIcon {
-  background-position: -158px -316px;
-  width: 32px;
-  height: 19px;
+  background-position: -185px -725px;
+  width: 27px;
+  height: 16px;
 }
 .icon.azureIcon {
-  background-position: -105px -724px;
-  width: 66px;
-  height: 19px;
+  background-position: -104px -725px;
+  width: 20px;
+  height: 16px;
 }
 .icon.googleCloudIcon {
-  background-position: -103px -748px;
-  width: 122px;
-  height: 19px;
+  background-position: -160px -725px;
+  width: 20px;
+  height: 16px;
 }
 .icon.cloudIcon {
-  background-position: -202px -313px;
-  width: 29px;
-  height: 19px;
+  background-position: -130px -725px;
+  width: 25px;
+  height: 16px;
 }
 .icon.blueWarningIcon {
   background-position: -157px -26px;
@@ -654,4 +654,9 @@
   background-position: -108px -289px;
   width: 13px;
   height: 11px;
+}
+.icon.external-link-icon {
+  background-position: -105px -749px;
+  width: 12px;
+  height: 12px;
 }
